# Conflicts:
#	gradle.properties

+++ src/main/java/dev/xkmc/youkaishomecoming/content/spell/spellcard/CardHolder.java
@@ -28,7 +28,7 @@ public interface CardHolder {
 
 	ItemDanmakuEntity prepareDanmaku(int life, Vec3 vec, YHDanmaku.Bullet type, DyeColor color);
 
-	ItemLaserEntity prepareLaser(int life, Vec3 pos, Vec3 vec, int len, YHDanmaku.Laser type, DyeColor color);
+	ItemLaserEntity prepareLaser(int life, Vec3 pos, Vec3 vec, float len, YHDanmaku.Laser type, DyeColor color);
 
 	void shoot(Entity danmaku);
